tambah peringatan ketika inputan null
inputan pada pembuatan pertanyaan dan jawaban lupa pin

+++ createpin.py
@@ -20,13 +20,17 @@ class BuatPINScreen(Screen):
             digitkurang.open()
             return 'buatpin'
         else:
-            if self.newpin.text == self.confirmpin.text:
+            if self.newpin.text == self.confirmpin.text and self.pertanyaan.text != '' and self.jawaban.text != '':
                 berhasil = MDDialog(text='PIN berhasil dibuat.',
                                     size_hint=(0.6,0.2))
                 berhasil.open()
                 db.deletepin() #menghapus pin sebelumnya jika sudah ada
                 db.addpin(self.newpin.text, self.pertanyaan.text, self.jawaban.text) #memasukkan pin ke database
                 return 'settings'
+            elif self.pertanyaan.text == '' or self.jawaban.text == '':
+                peringatan = MDDialog(text='Harap buat pertanyaan dan jawaban untuk lupa pin.',size_hint=(0.7,0.2))
+                peringatan.open()
+                return 'buatpin'
             else: #jika pin dan konfirmasinya tidak pas/sama
                 tidaksama = MDDialog(text='Harap masukkan angka yang sama saat konfirmasi.',
                                  size_hint=(0.7, 0.2))
